requirements.txt
# Core Python Libraries
streamlit==1.29.0
python-dotenv==1.0.0

# AI and Machine Learning
crewai==0.28.8
google-generativeai==0.3.2

# Web Requests and Data Handling
requests==2.31.0
json5==0.9.17

# Optional but Recommended
pydantic==2.6.1
typing-extensions==4.9.0

# Development and Debugging
# Note: These are optional and can be removed if not needed
black==24.1.1
pytest==8.0.0
mypy==1.8.0
